Create README.md
As an old and retired audio engineer, I have given myself a task to upgrade what were once very good speakers. To this end I purchase new drivers to replace the 40+ year old speakers. I erroneously assumed that simply replacing the drivers and keeping the existing crossovers would be a simple solution but that was not the case and I am not sure why this did not do the intended job. I tried a couple of low cost crossovers but apparently the levels were not balanced and some distortion was apparent in the crossover regions. So now I am going through the design of a new setup using an active amp tp power the woofers and a new passive filter for the midrange and tweeter drivers.